MLIBZ-1373 Remove TestDataStore and TestSQLiteCache files
