+++ UnitTestFramework/UnitTestFramework.csproj
@@ -61,9 +61,6 @@
     <Compile Include="..\UnitTests\TestSetup.cs">
       <Link>UnitTests\TestSetup.cs</Link>
     </Compile>
-    <Compile Include="..\UnitTests\TestSQLiteCache.cs">
-      <Link>UnitTests\TestSQLiteCache.cs</Link>
-    </Compile>
     <Compile Include="..\UnitTests\TestFiles\FlashCard.cs">
       <Link>UnitTests\TestFiles\FlashCard.cs</Link>
     </Compile>
